chore(chapter10): rename case05 files to case05_1 for clarity
