+++ chapter10/src/case05/case05_1.py
